requirements.txt
asgiref==3.3.0
astroid==2.4.2
Django==3.1.3
django-crispy-forms==1.9.2
django-leaflet==0.27.1
django-pwa==1.0.10
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
psycopg2==2.8.6
pylint==2.6.0
python-google-places==1.4.2
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.1
uWSGI==2.0.19.1
whitenoise==5.2.0
wrapt==1.12.1
